# perf: dont use `singledispatchmethod`