The decorator is expensive. We make O(n) calls for each document (we init child docs using same code)

### Before

```py
In [1]: doc =  frappe.get_doc('User', 'Administrator')

In [2]: roles = [d.as_dict() for d in doc.roles]

In [3]: %timeit doc.extend('roles', roles)
1.15 ms ± 14.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [4]: %timeit doc.extend('roles', roles)
1.15 ms ± 10.7 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

### After

```py
In [1]: doc =  frappe.get_doc('User', 'Administrator')

In [2]: roles = [d.as_dict() for d in doc.roles]

In [3]: %timeit doc.extend('roles', roles)
900 µs ± 2.53 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [4]: %timeit doc.extend('roles', roles)
897 µs ± 17.7 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```